+++ CHANGELOG.md
@@ -7,18 +7,19 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Fixed
 - Fixed that `BlobWriteParams`, `StringWriteParams`, and `ByteArrayWriteParams` didn't set their `WriteCommandType` given from the constructor.
+- Fixed naming of `GetOriginPrivateDirectoryAsync` and `IsSupportedAsync` to have _Async_ in name.
 
 ## [1.2.0] - 2022-10-07
 ### Added
 - Added `ArrayBufferAsync` method to `Blob` to read as byte array. By [@fixnil](https://github.com/fixnil).
 - Added public `CreateAsync` methods to `FileSystemHandle`, `FileSystemFileHandle`, and `FileSystemDirectoryHandle`.
 ### Fixed
-- Fixed naming of `QueryPermissionAsync` and `RequestPermissionAsync` to have _async_ in name.
+- Fixed naming of `QueryPermissionAsync` and `RequestPermissionAsync` to have _Async_ in name.
 - Fixed that there was an extra _in_ in the name of `OpenFilePickerOptionsStartInFileSystemHandle`.
 
 ## [1.1.0] - 2022-08-19
 ### Added
-- `FileSystemWritableStream` now extends the `Stream`.
+- `FileSystemWritableFileStream` now extends the `Stream`.
 
 ## [1.0.1] - 2022-07-12
 ### Fixed